Code Refactoring: The Great File Split Up!
Separated components into their own files. This has been overdue for a
while now, as things were getting a bit hard to manage within a single
file (though it would've been nice to still have been able to have had
everything in a single file for ease of deployment).

A few things have had to be shuffled around to make things work ok:
- BranchPane.TYPE_* defines have been moved to coreDefines::BranchType
- Project class init is now done in coreDefines so that everything can
access it, but this needs double checking to ensure that everything
will get to use the same instance. Perhaps this should be migrated to
the singleton-factory pattern instead!