== Skip-Lists: Part 1 ==
Implemented backend support for "skip-list". The purpose of this
functionality is to mask off some files from being accidentally
modified from DualitySVN, thus allowing "local-only changes" (i.e.
those that should never see the light of day in the official
repository) to be set up.

While this is kindof replicating change-sets in SVN, the advantage
here is that this is merely for your own usage, and not something that
will be stored anywhere near the SVN metadata, and potentially
committed/modified as a result of that.