"Ignore Externals" feature
This is just a quick hack, but when the "Ignore Externals" option in
the Tools menu is enabled, SVN Update will only update stuff in the
local repository and ignore linked 'externals'. Use this if you update
regularly and end up
having to wait a while on such repositories to update each time.

TODO:
- File I/O support
- Cleaner property binding implementation